fix: change order of product constructor parameters
- change product info substring length
