+++ src/main/java/vendingmachine/utils/ProductParser.java
@@ -1,7 +1,5 @@
 package vendingmachine.utils;
 
-import static vendingmachine.validator.ProductFormatValidator.validateFields;
-import static vendingmachine.validator.ProductFormatValidator.validateProductFormat;
 import static vendingmachine.validator.ProductFormatValidator.validateProductInfo;
 
 import java.util.Arrays;
@@ -18,7 +16,7 @@ public class ProductParser {
     private static final int PRODUCT_AMOUNT_INDEX = 1;
     private static final int PRODUCT_STOCK_INDEX = 2;
     private static final int INFO_START_INDEX = 1;
-    private static final Function<String, Integer> INFO_END_INDEX = s -> s.length() - 2;
+    private static final Function<String, Integer> INFO_END_INDEX = s -> s.length() - 1;
 
     public static List<Product> parse(String formattedInfo) throws IllegalArgumentException {
         return Arrays.stream(formattedInfo.split(INFOS_DELIMITER))
@@ -30,11 +28,9 @@ public class ProductParser {
         validateProductInfo(info);
         String[] fields = info.substring(INFO_START_INDEX, INFO_END_INDEX.apply(info))
             .split(INFO_DELIMITER);
-        validateProductFormat(info);
-        validateFields(fields);
         return new Product(fields[PRODUCT_NAME_INDEX],
-            Integer.parseInt(fields[PRODUCT_STOCK_INDEX]),
-            Integer.parseInt(fields[PRODUCT_AMOUNT_INDEX]));
+            Integer.parseInt(fields[PRODUCT_AMOUNT_INDEX]),
+            Integer.parseInt(fields[PRODUCT_STOCK_INDEX]));
     }
 
 }
